+++ packages/nodejs/package.json
@@ -23,6 +23,7 @@
     ],
     "devDependencies": {
         "@types/jest": "^26.0.23",
+        "@types/google-protobuf": "^3.15.3",
         "@typescript-eslint/eslint-plugin": "^4.28.1",
         "@typescript-eslint/parser": "^4.28.1",
         "babel-jest": "^27.0.6",
@@ -30,6 +31,8 @@
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-import": "^2.23.4",
         "eslint-plugin-prettier": "^3.4.0",
+        "grpc-tools": "^1.11.2",
+        "grpc_tools_node_protoc_ts": "5.3.0",
         "jest": "^27.0.6",
         "lint-staged": "^12.0.2",
         "noble-ed25519": "^1.2.5",
@@ -43,10 +46,14 @@
         "tabWidth": 4
     },
     "scripts": {
+        "generate-js": "yarn run grpc_tools_node_protoc --js_out=import_style=commonjs,binary:grpc --grpc_out=grpc_js:grpc --plugin=protoc-gen-grpc=./node_modules/.bin/grpc_tools_node_protoc_plugin -I ../../deps/concordium-grpc-api ../../deps/concordium-grpc-api/*.proto",
+        "generate-ts": "yarn run grpc_tools_node_protoc --plugin=protoc-gen-ts=./node_modules/.bin/protoc-gen-ts --ts_out=grpc_js:grpc -I ../../deps/concordium-grpc-api ../../deps/concordium-grpc-api/*.proto",
+        "generate": "mkdir -p grpc && yarn generate-js && yarn generate-ts",
         "lint": "eslint . --cache --ext .ts,.tsx --max-warnings 0",
         "lint-fix": "yarn --silent lint --fix; exit 0",
         "test": "jest",
-        "build": "tsc"
+        "prebuild": "if [ ! -e \"grpc\" ]; then yarn generate; fi",
+        "build": "tsc && cp -r ./grpc/ ./lib/ "
     },
     "dependencies": {
         "@grpc/grpc-js": "^1.3.4",
